requirements.txt
streamlit
ollama
pymupdf
sentence-transformers
transformers
torch
faiss-cpu
pillow
numpy
